requirements.txt
setuptools
aiohttp
tornado
strsimpy
cpca
pymysql
pytest
pytest-asyncio
opencc
dimsim
rasa-nlu==0.12.2
mitie==0.7.36
jieba==0.42.1
numpy==1.16.5
scikit-learn==0.19.2
scipy==1.6.2
spacy==3.0.5
spacy-pkuseg==0.0.28
pyunit-time==2021.2.2
pypinyin==0.41.0
ngram==3.3.2
paddlepaddle-tiny==1.6.1
cn2an==0.5.11
pyyaml==5.4.1
